[REFACTOR] Separating Step and Dialog components into separate files.
